+++ README.md
@@ -61,3 +61,4 @@
 * Error가 응답 객체를 반환하도록 변경
 * 인터셉터 중복 제거
 * 1차 피드백 반영
+* AdminController 분리
