+++ src/main/java/nextstep/RoomEscapeApplication.java
@@ -2,10 +2,12 @@ package nextstep;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.http.HttpMethod;
 
 @SpringBootApplication
 public class RoomEscapeApplication {
     public static void main(String[] args) {
+        HttpMethod.POST.name();
         SpringApplication.run(RoomEscapeApplication.class, args);
     }
 }
