+++ src/main/java/nextstep/interfaces/controller/AdminController.java
@@ -1,0 +1,31 @@
+package nextstep.interfaces.controller;
+
+import lombok.RequiredArgsConstructor;
+import nextstep.domain.model.request.ScheduleRequest;
+import nextstep.domain.model.request.ThemeRequest;
+import nextstep.service.ScheduleService;
+import nextstep.service.ThemeService;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.net.URI;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/admin/themes")
+public class AdminController {
+    private final ThemeService themeService;
+    @PostMapping
+    public ResponseEntity<Void> createTheme(@RequestBody ThemeRequest themeRequest) {
+        Long id = themeService.create(themeRequest);
+        return ResponseEntity.created(URI.create("/themes/" + id)).build();
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteTheme(@PathVariable Long id) {
+        themeService.delete(id);
+
+        return ResponseEntity.noContent().build();
+    }
+}
+
